+++ services/app.module.ts
@@ -1,6 +1,6 @@
 import { Module } from "@nestjs/common";
 import { OrganizationsModule } from "./organizations/organizations.module";
-import { PeruConnectModule } from "./peru-connect/peru-connect.module";
+import { PeruConnectModule } from "./sunat/sunat.module";
 import { UsersModule } from "./users/users.module";
 import { AuthModule } from "./auth/auth.module";
 import { SecurityModule } from "./security/security.module";
