+++ services/sunat/encore.service.ts
@@ -1,3 +1,3 @@
 import { Service } from "encore.dev/service";
 
-export default new Service("peru_connect");
+export default new Service("sunat");
